requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-filter==1.1.0
django-oauth-toolkit==1.0.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
Markdown==2.6.9
oauthlib==2.0.6
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
shortuuid==0.5.0
urllib3==1.22
